
\name{cult_change_for_predictiveCheck}
\alias{cult_change_for_predictiveCheck}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{}
\description{

} 
\usage{
cult_change_for_predictiveCheck<-function(pop)
}
\arguments{
    \item{pop}{population structure at time t1}
}

\details{
}
%- maybe also 'usage' for other objects documented here.
\references{}
\author{Clementine Straub}
\examples{

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ }